Updated embolden_home.js
Spent a little bit of time working on the code of this Home screen. I made styles for the Title and Subtext and also tried to figure out to flex the ".j." to the bottom of the app.

I'm going to revisit this home screen in a couple of tutorials once I work through "react-native-navigation" and also get a better feel for how I want this app to be.

Thanks for reading! Goodnight 😴 
.j.