Created the BoldButton .j.
Hello GitHub! For the Embolden app, I had an idea to make a simple "Bold Button" component that a missionary can press and it will pop up Bible verses to encourage them in the button.

Made a simple Text component that has an onPress function that pops up a simple alert. Going to add more logic to it to pool from a list of verses and stuff like that!

Thank you for reading!
.j.