Embolden | Expo Redesign
Hello GitHub! Following the wonderful Expo Create App tutorials (https://docs.expo.dev/tutorial), I recreated the barebones of this app in Expo!

I have started to storyboard the app and added in tabs for each of the different screens that I want in the app. Now I am going to work on adding in logic, functionality, finishing the Figma storyboard, and working on branding and colors.

There are still a lot of things I need to work on for this app, but I am happy to say that I redesigned the app in Expo! Excited to learn and keep creating!

Thank you for reading!
.j.