Made a BoldButton component
Hello GitHub! Tonight I added in my old BoldButton.js code and added it as a file in the components folder! I imported it into the index.js file to make sure it works and it works. Also changed the color and the style for now and created a const that links to a "encouragingVersesArray." A new goal is to make the BoldButton verses randomly generated.

Thank you for reading!
.j.