Updated index.js
Did a good bit of work on this one! Added in state so I could make the new login modal appear and disappear with a button (still need to fine tune this) as well as restyled the actual content and styling of the title, subtitle and UI layout of this page.

Liked working on this tonight!
.j.