Updated index.js and added Poppins-Black
Spent some time working on Embolden tonight after going through a tutorial and got some decent work done!

Took out ScrollView, added in TouchableOpacity, reworked the code and where it is and laid out the Views so that I could try to get the BOLD button to flex to the bottom. Did the same with the login button (More on that code in other commits). Changed some styling as well on the page and added in Poppins-Black.

Had a good time coding tonight!
.j.